pyproject.toml
[tool.poetry]
name = "food-tracker"
version = "0.1.0"
description = "A system to track eating habits by analyzing food photos"
authors = ["Dipam Vasani"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pillow = "^10.1.0"
python-dotenv = "^1.0.0"
instructor = "^0.5.0"
openai = "^1.14.0"
anthropic = "^0.19.0"
google-api-python-client = "^2.118.0"
google-auth-oauthlib = "^1.2.0"
pydantic = "^2.6.0"
tqdm = "^4.66.1"
matplotlib = "^3.8.3"
numpy = "^1.26.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 88
